splitting up the blockstore file
